Simplified pay_item as the current intention is that payslip manages periodivity. pay_item can specify whether it is one off.
